Updated Rector to commit 9a0938d8df9d8771d8d423eb4719cd7fe3b1ba0e
https://github.com/rectorphp/rector-src/commit/9a0938d8df9d8771d8d423eb4719cd7fe3b1ba0e [e2e] handle multiple classes with anonymous class on MultipleClassFileToPsr4ClassesRector file removed (#3253)

+++ rules/PSR4/Rector/Namespace_/MultipleClassFileToPsr4ClassesRector.php
@@ -8,9 +8,11 @@ use PhpParser\Node\Stmt\Class_;
 use PhpParser\Node\Stmt\ClassLike;
 use PhpParser\Node\Stmt\Namespace_;
 use Rector\Core\Application\FileSystem\RemovedAndAddedFilesCollector;
+use Rector\Core\NodeAnalyzer\ClassAnalyzer;
 use Rector\Core\PhpParser\Node\CustomNode\FileWithoutNamespace;
 use Rector\Core\PhpParser\Printer\NeighbourClassLikePrinter;
 use Rector\Core\Rector\AbstractRector;
+use Rector\FileSystemRector\ValueObject\AddedFileWithContent;
 use Rector\PSR4\FileInfoAnalyzer\FileInfoDeletionAnalyzer;
 use Rector\PSR4\NodeManipulator\NamespaceManipulator;
 use Symplify\RuleDocGenerator\ValueObject\CodeSample\CodeSample;
@@ -40,12 +42,18 @@ final class MultipleClassFileToPsr4ClassesRector extends AbstractRector
      * @var \Rector\Core\Application\FileSystem\RemovedAndAddedFilesCollector
      */
     private $removedAndAddedFilesCollector;
-    public function __construct(NamespaceManipulator $namespaceManipulator, FileInfoDeletionAnalyzer $fileInfoDeletionAnalyzer, NeighbourClassLikePrinter $neighbourClassLikePrinter, RemovedAndAddedFilesCollector $removedAndAddedFilesCollector)
+    /**
+     * @readonly
+     * @var \Rector\Core\NodeAnalyzer\ClassAnalyzer
+     */
+    private $classAnalyzer;
+    public function __construct(NamespaceManipulator $namespaceManipulator, FileInfoDeletionAnalyzer $fileInfoDeletionAnalyzer, NeighbourClassLikePrinter $neighbourClassLikePrinter, RemovedAndAddedFilesCollector $removedAndAddedFilesCollector, ClassAnalyzer $classAnalyzer)
     {
         $this->namespaceManipulator = $namespaceManipulator;
         $this->fileInfoDeletionAnalyzer = $fileInfoDeletionAnalyzer;
         $this->neighbourClassLikePrinter = $neighbourClassLikePrinter;
         $this->removedAndAddedFilesCollector = $removedAndAddedFilesCollector;
+        $this->classAnalyzer = $classAnalyzer;
     }
     public function getRuleDefinition() : RuleDefinition
     {
@@ -109,9 +117,14 @@ CODE_SAMPLE
         if ($nodeToReturn !== null) {
             return $nodeToReturn;
         }
-        // 2. nothing to return - remove the file
-        $this->removedAndAddedFilesCollector->removeFile($this->file->getFilePath());
-        return null;
+        $isInaddedFiles = \array_filter($this->removedAndAddedFilesCollector->getAddedFilesWithContent(), function (AddedFileWithContent $addedFileWithContent) : bool {
+            return $addedFileWithContent->getFilePath() === $this->file->getFilePath();
+        });
+        if ($isInaddedFiles === []) {
+            // 2. nothing to return - remove the file
+            $this->removedAndAddedFilesCollector->removeFile($this->file->getFilePath());
+        }
+        return $node;
     }
     private function hasAtLeastTwoClassLikes(Node $node) : bool
     {
@@ -165,11 +178,11 @@ CODE_SAMPLE
     private function findNonAnonymousClassLikes(Node $node) : array
     {
         $classLikes = $this->betterNodeFinder->findInstanceOf([$node], ClassLike::class);
-        return \array_filter($classLikes, static function (ClassLike $classLike) : bool {
+        return \array_filter($classLikes, function (ClassLike $classLike) : bool {
             if (!$classLike instanceof Class_) {
                 return \true;
             }
-            return !$classLike->isAnonymous();
+            return !$this->classAnalyzer->isAnonymousClass($classLike);
         });
     }
 }
